@@ -34,13 +34,14 @@
     "passport": "^0.3.2",
     "passport-local": "^1.0.0",
     "passport-local-mongoose": "^4.0.0",
+    "prop-types": "^15.5.10",
     "pug": "^2.0.0-beta11",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "react-redux": "^5.0.3",
-    "redux": "^3.6.0",
-    "redux-devtools": "^3.3.2",
-    "redux-saga": "^0.15.3",
+    "react": "^15.6.1",
+    "react-dom": "^15.6.1",
+    "react-redux": "^5.0.5",
+    "redux": "^3.7.1",
+    "redux-devtools": "^3.4.0",
+    "redux-saga": "^0.15.4",
     "semantic-ui-checkbox": "^2.2.9",
     "semantic-ui-dimmer": "^2.2.9",
     "semantic-ui-dropdown": "^2.2.9",
@@ -80,7 +81,6 @@
     "moment": "^2.18.1",
     "node-notifier": "^5.1.2",
     "rc-slider": "5.1.3",
-    "react-test-renderer": "^15.5.4",
     "require-dot-file": "^0.4.0",
     "through2": "^2.0.3",
     "vinyl-transform": "^1.0.0",
