@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
 export default class Changelog extends React.Component {
 	constructor() {
@@ -18,58 +19,18 @@ export default class Changelog extends React.Component {
 					<h2>Changelog</h2>
 				</div>
 				<div className="ui header">
-					<p>Version 0.6.3 "conte" released 6-24-2017</p>
+					<p>Version 0.6.4 "dry brush" released 7-9-2017</p>
 				</div>
-				<h3>New feature: game replays.</h3>
-				<img style={{width: '100%'}} src="http://i.imgur.com/7ANlC7L.gif" />
-				<p>After a game, click on the new button next to leave game to check them out, or in your profile.</p>
-				<h3>New feature: moderator IP bans.</h3>
-				<p>Obviously need to be used carefully, but this will prevent (hopefully) some of the worst offenders from repeatedly making accounts. All IP bans are temporary and go away relatively fast.</p>
+				<h3>New feature: player color refactoring.</h3>
+				<p>Player colors have been changed in two ways: they are fixed by math as opposed to me randomly clicking colors, and there are many more shades (10) that change every 2% points for purples.  Also, purple starts at 52% as opposed to 55%. Check the "!" icon next to the lobby.</p>
+				<h3>New feature: player color on card in game.</h3>
 				<p>Other issues:</p>
 				<ul>
-					<li>Observer chat is back.</li>
-					<li>A fix to non rainbows having purple names in the userlist/lobby.</li>
-					<li>A fix to clicking on an icon in the userlist/lobby taking you to the profile instead of the game.</li>
-					<li>A fix to a general styling issue - seen most obviously by the yellow circle around the lizard being offset.</li>
-					<li>A fix to only rainbow players in a rainbow game showing up on the userlist/lobby as opposed to all rainbow players.</li>
-					<li>Thanks to a submission, tracks have been manually desaturated and replaced, most users won't see a change.</li>
-					<li>The "delete started game" cleanup feature has been moved from 30 minutes after the game starts to 40 minutes, per request.</li>
-					<li>All images minified so should be a lot snappier on first load..</li>
+					<li>There is one crash bug left that has happened once in the past 10 days and hopefully I've prevented it permanetly in this patch.</li>
+					<li>Some fixes to bugs in replays, and replays now have vetos.</li>
+					<li>The rules page for this site has been adjusted to account for online vs physical play.</li>
+					<li>Private games now have a new symbol in the gamelist as opposed to green text.</li>
 				</ul>
-				<h3>New feature: Replays</h3>
-				<img src="/images/replay-demo.gif" width="750px" />
-				<ul>
-					<li>Replays are now available! They can be accessed from the post-game or from player profiles.</li>
-				</ul>
-				<h3>Misc. bug fixes</h3>
-				<ul>
-					<li>Fixed bug with Vote Accuracy, which was counting elections with Liberal presidents.</li>
-					<li>Fixed bug that was sometimes awarding false losses to profiles after a Liberal win.</li>
-				</ul>
-				<div className="ui header">
-					<p>Version 0.6.2 "ssalg deniats" released 6-18-2017</p>
-				</div>
-				<ul>
-					<li>A fix to the userlist display issues.</li>
-					<li>A fix to non rainbows having purple names and colors.</li>
-					<li>Back end validation to chats, meaning dead players can't avoid the disabled filter & not logged in users can't chat.</li>
-					<li>Player names are now case insensitive, meaning that a player cannot make an account named "Valeera" if there's already an account called "valeera". If some accounts need to be deleted let me know.</li>
-				</ul>
-				<div className="ui header">
-					<p>Version 0.6.1 "stained glass" released 6-17-2017</p>
-				</div>
-				<h3>New feature: player font size setting in game settings.</h3>
-				<p>Click the gear icon to change the size of the font in games.</p>
-				<h3>New feature: admin and mod broadcasts.</h3>
-				<p>Admins and mods have the ability to send a message to all games.  This should be used sparingly and really only for downtime announcements or if a particularly bad troll attack happens.</p>
-				<h4>Other updates</h4>
-				<ul>
-					<li>A fix to the general styling problems with profiles.</li>
-					<li>A fix to player's win rates being bumped down a line due to long usernames/icons.</li>
-					<li>Rainbow player's lobby is now sorted correctly.</li>
-				</ul>
-				<p>New morning mods: Faaiz1999 and DumbBullDoor.  We have pretty good mod coverage so closing mod applications for a bit, will reopen when needed.  Thanks!</p>
-				<p>Next up: more bug fixes in preparation for a 1.0 beta and release (all bugs fixed, all desired features in).</p>
 				<div className="ui header">
 					<p>Version 0.6.0 "noise" released 6-12-2017</p>
 				</div>
@@ -116,5 +77,5 @@ export default class Changelog extends React.Component {
 }
 
 Changelog.propTypes = {
-	onLeaveChangelog: React.PropTypes.func,
+	onLeaveChangelog: PropTypes.func,
 };
